backend/requirements.txt
Django==5.0

django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-registration==0.1.3
drf-yasg==1.21.7
swagger-spec-validator==3.0.3

daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
